refactor(receiver-mock): move terraform handlers to a separate module
